requirements.txt
pygame
python-chess
pygame-textinput
google-generativeai
